therapy_buddy_backend/requirement.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.30.1
python-dotenv==1.0.1
pydantic[email]==2.7.1
httpx==0.27.0
python-multipart==0.0.9